feat: 重构dts;新增HTMLBuild

+++ src/createExportFile.ts
@@ -1,10 +1,6 @@
-import {
-  SrcModuleInfo,
-  windowsPathToLinuxPath,
-  getRootDirname,
-} from "module-ctrl";
+import { SrcModuleInfo, formatLinuxPath, getRootDirname } from "module-ctrl";
 import { wirteJsonFile } from "./utils/JsonFile";
-import { promises as fs } from "fs";
+import { promises as fs, existsSync } from "fs";
 import path from "path";
 import { Exception } from "exception";
 
@@ -23,30 +19,33 @@ export async function createExportFile(fileName: string): Promise<void> {
 
   // 使用 path 拼接路径
   const filePath = path.join(rootDir, srcDir, `${fileName}.ts`);
-  const relativePath = windowsPathToLinuxPath(
+  const relativePath = formatLinuxPath(
     path.join("./", outputDir, `${fileName}${dist ? ".js" : ".ts"}`),
     true
   );
 
   // 创建文件
-  await fs.writeFile(filePath, "", "utf-8");
+  if (!existsSync(filePath)) {
+    await fs.writeFile(filePath, "", "utf-8");
+  }
 
+  const rname = formatLinuxPath(fileName, true);
   if (dist) {
     packageJson.srcModule.dist = packageJson.srcModule.dist || {};
-    packageJson.srcModule.dist[fileName] = windowsPathToLinuxPath(
+    packageJson.srcModule.dist[rname] = formatLinuxPath(
       path.join(srcDir, `${fileName}.ts`),
       true
     );
-    packageJson.exports[fileName] = {
+    packageJson.exports[rname] = {
       import: relativePath,
       require: relativePath.replace(".js", ".cjs"),
       types: relativePath.replace(".js", ".d.ts"),
     };
   } else {
-    packageJson.exports[fileName] = relativePath;
+    packageJson.exports[rname] = relativePath;
   }
 
   // 保存更新后的 package.json
   await wirteJsonFile(path.join(rootDir, "package.json"), packageJson, 2);
-  console.log(`创建文件 ${filePath} 成功`);
+  console.log(`创建 ${filePath} 成功`);
 }
